Fixing some errors
I forgot that Octave's return just exits the function; I have to specify the return values when declaring the function, kind of like Fortran. I have done so and used this opportunity to fix other minor errors that happened at execution. I also moved the `random_choice` function to a separate folder, just so that it is not exposed when someone else calls my functions. With this, I should be able to move towards converting the third file from the original project, which computes the arrivable graph.