Back to loops
I tried a little to hard in vectorizing these functions when speed is not a requirement as of now; ensuring that they are working correctly takes precedence. This means that its back to normal loops for the time being and also some additional comments for clarity, although I could probably improve these, too.