Generating the Reachability Graph
This is one of the more important functions for the process, since we use the reachability graph when generating the performance metrics. I should probably change the function name so that it is closer to the actual name of the operation, but I could do this later. I should review this code carefully, since there are a few confusing things on the original implementation, specially regarding the iteration over the enabled markings and the dot operation used to create a new marking; I feel like this is not actually a dot, since we are using a matrix and a vector.