SPN matrix type
For the initial SPN, a boolean matrix is enough, but this will certainly lead to overflows when we are verifying the boundedness and generating the reachability tree. For the time being, `uint8` is going to be the default value type used, but using something like `uint16` or `uint32` would be a better idea. I am fairly certain that generating wider integers is going to be a better solution than to manually widen it later on.