Formatting, comments and a minor fix
I removed some line breask since 80 characters is pretty restrictive, so
this should help a bit with parsing; I will fix the comments to use this
additional line space later on. Other than that, there was a slight
mistake on the `enalbed_stets` function, where I was using a
non-existent matrix instead of the correct A1; this has now been fixed.