Select colums by another name
The dot call being used in the original Python code is nothing more than
a row selector in the end, but, since I still want to keep this close, I
will be keeping the actual operation being performed: a sum product over
the last axis of the matrix and vector. On the next commit, I will
change this to the proper selection operation.