Minor corrections
I was doing a bunch of sums when simple comparisons would work, so I
have changed the functions in order to do just that. This shouldn't make
much of a difference in terms of performance, but it does get easier to
follow, in my opinion. Likewise, there are some additional comments for
further clarity.