Default data type confusion
Since there is no way of knowing beforehand if the generated Petri nets
are going to be bounded, and even if they are bounded at low values, it
makes little sense to save bits at this starting point. I could also
just keep the previous value for ints and then increase their size when
solving the nets, so this could be reveresed later.