Better comments, rewrite of add_node
There were some points of improvement on the add_node function, so
I rewrote it in a more compact form which is still readable: this
should be easier to parse visually and also more performant. I do
need to run some tests on this function to ensure that it retains
its original funcionality.